requirement.txt
tensorflow
tensorflow-hub
pillow
pymodbustcp
pyserial
wheel
opencv-contrib-python
numpy
matplotlib
scipy
pylibdmtx
watchdog
pyzbar
opencv_python
scikit_image
scikit_learn
torch
torchvision
tqdm
git+https://github.com/OpenAOI/anodet.git
pytesseract
open3d
pandas
pyvista
joblib
xmltodict
easyocr